cleanup

+++ src/Engine/OcclusionCulling.cpp
@@ -72,8 +72,8 @@ void OcclusionCuller::rasterizeOccluders(std::vector<Object*> objs)
 {
 	auto* buff = raster.getBuff();	
 
-	for (size_t i = 0; i < raster.getW(); ++i) {
-		for(size_t j = 0; j < raster.getH(); ++j){
+	for (uint32_t i = 0; i < raster.getW(); ++i) {
+		for(uint32_t j = 0; j < raster.getH(); ++j){
 			*raster.at(i, j) = 128;//(i + j) > 255 ? 6 : (i + j);
 		}
 	}
